+++ plone/app/testing/cleanup.py
@@ -13,7 +13,7 @@ except ImportError:
 
 # Make sure cleanup handlers from PAS are registered
 try:
-    import Products.PluggableAuthService.zcml
+    import Products.PluggableAuthService.zcml  # noqa
 except ImportError:
     pass