+++ CHANGES.rst
@@ -1,6 +1,18 @@
 Changelog
 =========
 
+5.0.1 (unreleased)
+------------------
+
+New:
+
+- *add item here*
+
+Fixes:
+
+- *add item here*
+
+
 5.0.0 (2016-02-20)
 ------------------